% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMCFB-package.R
\docType{package}
\name{DMCFB-package}
\alias{DMCFB-package}
\alias{DMCFB}
\title{Differentially Methylated cytosines using funtional Bayesian
regression models}
\description{
DMCFB is a profiling tool for identifying differentially
methylated cytosines using Functional Bayesian Model in bisulfite sequencing
data.
}
\section{\code{DMCFB} methods}{

\code{\link{findDMCFB}},
\code{\link{plotDMCFB}},
\code{\link{cBSDMC}},
\code{\link{readBismark}}.
}

\section{\code{BSDMC} objects}{

\code{\link{BSDMC-class}}
}

