% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/writeBED-method.R
\docType{methods}
\name{writeBED-method}
\alias{writeBED-method}
\alias{writeBED,BSDMC,character,missing-method}
\alias{writeBED}
\alias{writeBED,BSDMC,missing,character-method}
\alias{writeBED,BSDMC,missing,missing-method}
\title{writeBED method}
\usage{
writeBED(object, name, file)

\S4method{writeBED}{BSDMC,character,missing}(object, name)

\S4method{writeBED}{BSDMC,missing,character}(object, file)

\S4method{writeBED}{BSDMC,missing,missing}(object)
}
\arguments{
\item{object}{A \code{\link{BSDMC-class}}
object}

\item{name}{A character list}

\item{file}{A character}
}
\value{
BED files
}
\description{
write BS-Seq data to BED files
}
\examples{
fn <- list.files(system.file("extdata", package = "DMCHMM"))
fn.f <- list.files(system.file("extdata", package = "DMCHMM"),
  full.names = TRUE
)
OBJ <- readBismark(fn.f, fn)
cdOBJ <- DataFrame(Cell = factor(c("BC", "TC", "Mono"),
  labels = c("BC", "TC", "Mono")
), row.names = c("BCU1568", "BCU173", "BCU551"))
colData(OBJ) <- cdOBJ
writeBED(OBJ)
}
\author{
Farhad Shokoohi <shokoohi@icloud.com>
}
